requirement.txt
Package          Version  
---------------- ---------
absl-py          0.2.2    
astor            0.6.2    
backcall         0.1.0    
bleach           1.5.0    
certifi          2018.4.16
chardet          3.0.4    
decorator        4.3.0    
gast             0.2.0    
grpcio           1.12.1   
gspread          3.0.0    
h5py             2.8.0    
html5lib         0.9999999
httplib2         0.11.3   
idna             2.7      
ipdb             0.11     
ipython          6.4.0    
ipython-genutils 0.2.0    
jedi             0.12.0   
Keras            2.1.5    
Markdown         2.6.11   
nltk             3.3      
numpy            1.14.5   
oauth2client     4.1.2    
pandas           0.23.1   
parso            0.2.1    
pexpect          4.6.0    
pickleshare      0.7.4    
pip              10.0.1   
prompt-toolkit   1.0.15   
protobuf         3.6.0    
ptyprocess       0.5.2    
pyasn1           0.4.3    
pyasn1-modules   0.2.1    
Pygments         2.2.0    
python-dateutil  2.7.3    
pytz             2018.4   
PyYAML           3.12     
requests         2.19.1   
rsa              3.4.2    
scikit-learn     0.19.1   
scipy            1.1.0    
setuptools       39.2.0   
simplegeneric    0.8.1    
six              1.11.0   
sklearn          0.0      
tensorboard      1.8.0    
tensorflow-gpu   1.0.0    
termcolor        1.1.0    
traitlets        4.3.2    
urllib3          1.23     
wcwidth          0.1.7    
Werkzeug         0.14.1   
wheel            0.31.1   
xgboost          0.72     
